+++ controller/konnect/errors.go
@@ -2,6 +2,7 @@ package konnect
 
 import (
 	"fmt"
+	"time"
 
 	"k8s.io/apimachinery/pkg/types"
 )
@@ -34,4 +35,29 @@ type ReferencedKongServiceIsBeingDeleted struct {
 // Error implements the error interface.
 func (e ReferencedKongServiceIsBeingDeleted) Error() string {
 	return fmt.Sprintf("referenced Kong Service %s is being deleted", e.Reference)
+}
+
+// ReferencedKongConsumerIsBeingDeleted is an error type that is returned when
+// a Konnect entity references a Kong Consumer which is being deleted.
+type ReferencedKongConsumerIsBeingDeleted struct {
+	Reference         types.NamespacedName
+	DeletionTimestamp time.Time
+}
+
+// Error implements the error interface.
+func (e ReferencedKongConsumerIsBeingDeleted) Error() string {
+	return fmt.Sprintf("referenced Kong Consumer %s is being deleted (deletion timestamp: %s)",
+		e.Reference, e.DeletionTimestamp,
+	)
+}
+
+// ReferencedKongConsumerDoesNotExist is an error type that is returned when the referenced KongConsumer does not exist.
+type ReferencedKongConsumerDoesNotExist struct {
+	Reference types.NamespacedName
+	Err       error
+}
+
+// Error implements the error interface.
+func (e ReferencedKongConsumerDoesNotExist) Error() string {
+	return fmt.Sprintf("referenced Kong Consumer %s does not exist: %v", e.Reference, e.Err)
 }