+++ controller/konnect/index.go
@@ -21,9 +21,11 @@ func ReconciliationIndexOptionsForEntity[
 	T constraints.SupportedKonnectEntityType,
 ]() []ReconciliationIndexOption {
 	var e TEnt
-	switch any(e).(type) { //nolint:gocritic // TODO: add index options required for other entities
+	switch any(e).(type) {
 	case *configurationv1alpha1.KongPluginBinding:
 		return IndexOptionsForKongPluginBinding()
+	case *configurationv1alpha1.CredentialBasicAuth:
+		return IndexOptionsForCredentialsBasicAuth()
 	}
 	return nil
 }