+++ test/envtest/deploy_resources.go
@@ -7,11 +7,13 @@ import (
 	"github.com/google/uuid"
 	"github.com/samber/lo"
 	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/kong/gateway-operator/controller/konnect/conditions"
 
+	configurationv1 "github.com/kong/kubernetes-configuration/api/configuration/v1"
 	configurationv1alpha1 "github.com/kong/kubernetes-configuration/api/configuration/v1alpha1"
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
@@ -142,6 +144,33 @@ func deployKongService(
 	return kongService
 }
 
+// deployKongConsumerWithProgrammed deploys a KongConsumer resource and returns the resource.
+func deployKongConsumerWithProgrammed(
+	t *testing.T,
+	ctx context.Context,
+	cl client.Client,
+	consumer *configurationv1.KongConsumer,
+) *configurationv1.KongConsumer {
+	t.Helper()
+
+	consumer.GenerateName = "kongconsumer-"
+	require.NoError(t, cl.Create(ctx, consumer))
+	t.Logf("deployed %s KongConsumer resource", client.ObjectKeyFromObject(consumer))
+
+	consumer.Status.Conditions = []metav1.Condition{
+		{
+			Type:               conditions.KonnectEntityProgrammedConditionType,
+			Status:             metav1.ConditionTrue,
+			Reason:             conditions.KonnectEntityProgrammedReasonProgrammed,
+			ObservedGeneration: consumer.GetGeneration(),
+			LastTransitionTime: metav1.Now(),
+		},
+	}
+	require.NoError(t, cl.Status().Update(ctx, consumer))
+
+	return consumer
+}
+
 // deployKongPluginBinding deploys a KongPluginBinding resource and returns the resource.
 // The caller can also specify the status which will be updated on the resource.
 func deployKongPluginBinding(
@@ -158,4 +187,36 @@ func deployKongPluginBinding(
 
 	require.NoError(t, cl.Status().Update(ctx, kpb))
 	return kpb
+}
+
+// deployCredentialBasicAuth deploys a CredentialBasicAuth resource and returns the resource.
+func deployCredentialBasicAuth(
+	t *testing.T,
+	ctx context.Context,
+	cl client.Client,
+	consumerName string,
+	username string,
+	password string,
+) *configurationv1alpha1.CredentialBasicAuth {
+	t.Helper()
+
+	c := &configurationv1alpha1.CredentialBasicAuth{
+		ObjectMeta: metav1.ObjectMeta{
+			GenerateName: "basic-auth-",
+		},
+		Spec: configurationv1alpha1.CredentialBasicAuthSpec{
+			ConsumerRef: corev1.LocalObjectReference{
+				Name: consumerName,
+			},
+			CredentialBasicAuthAPISpec: configurationv1alpha1.CredentialBasicAuthAPISpec{
+				Password: password,
+				Username: username,
+			},
+		},
+	}
+
+	require.NoError(t, cl.Create(ctx, c))
+	t.Logf("deployed new unmanaged CredentialBasicAuth %s", client.ObjectKeyFromObject(c))
+
+	return c
 }